Crate pallet_nft_sales

NFT Sales pallet

This pallet provides a place for digital art creators and owners to offer their NFTs for sale and for potential buyers to browse and buy them.

To sell NFTs, users will call Pallet::add(). Doing so will have the NFT being transferred from the seller to this pallet’s account, to grant us the own

To remove an NFT from sale and thus reclaim its ownership, sellers will call Pallet::remove().

To buy an NFT, users will call buy.

Re-exports

pub use weights::WeightInfo;

pub use pallet::*;

Modules

pallet

The module that hosts all the FRAME types needed to add this pallet to a runtime.

weights

Structs

Price

Sale