Type Alias pallet_nft_sales::pallet::Module

pub type Module<T> = Pallet<T>;

👎Deprecated: use Pallet instead

Type alias to Pallet, to be used by construct_runtime.

Generated by pallet attribute macro.

Aliased Type

struct Module<T>(/* private fields */);

Implementations

impl<T: Config> Pallet<T>

pub fn add( origin: OriginFor<T>, class_id: T::CollectionId, instance_id: T::ItemId, price: Price<<<T as Config>::Fungibles as Inspect<<T as Config>::AccountId>>::AssetId, <<T as Config>::Fungibles as Inspect<<T as Config>::AccountId>>::Balance> ) -> DispatchResult

Add an NFT

Fails if

the NFT is not found in Config::NonFungibles

origin is not the owner of the nft

the nft is already for sale

transferring ownership of the NFT to this pallet’s account fails

pub fn remove( origin: OriginFor<T>, class_id: T::CollectionId, instance_id: T::ItemId ) -> DispatchResult

Remove an NFT

The seller of an NFT that is for sale can call this extrinsic to reclaim ownership over their NFT and remove it from sale.

Fails if

the nft is not for sale

origin is not the seller of the NFT

transferring the ownership of the NFT back to the seller fails

pub fn buy( origin: OriginFor<T>, class_id: T::CollectionId, instance_id: T::ItemId, max_offer: Price<<<T as Config>::Fungibles as Inspect<<T as Config>::AccountId>>::AssetId, <<T as Config>::Fungibles as Inspect<<T as Config>::AccountId>>::Balance> ) -> DispatchResult

Buy the given nft

Buyers must propose a max_offer to save them from a scenario where they could end up paying more than they desired for an NFT. That scenario could take place if the seller increased the asking price right before the buyer submits this call to buy said NFT.

Buyers always pay the latest asking price as long as it does not exceed their max offer.

Fails if

the NFT is not for sale

origin is the seller of the NFT

origin does not have enough balance of the currency the nft is being sold in

transferring the asking price from the buyer to the seller fails

transferring the nft to the buyer fails

impl<T: Config> Pallet<T>

pub fn origin() -> T::RuntimeOrigin

pub fn account() -> T::AccountId

impl<T: Config> Pallet<T>

pub fn get_sale<KArg1, KArg2>( k1: KArg1, k2: KArg2 ) -> Option<Sale<<T as Config>::AccountId, <<T as Config>::Fungibles as Inspect<<T as Config>::AccountId>>::AssetId, <<T as Config>::Fungibles as Inspect<<T as Config>::AccountId>>::Balance>>where KArg1: EncodeLike<T::CollectionId>, KArg2: EncodeLike<T::ItemId>,

The active sales. A sale is an entry identified by an NFT collection and item id.