Trait pallet_nft_sales::pallet::Config

pub trait Config: Config {
    type RuntimeEvent: From<Event<Self>> + IsType<<Self as Config>::RuntimeEvent>;
    type WeightInfo: WeightInfo;
    type Fungibles: Mutate<Self::AccountId>;
    type NonFungibles: Transfer<Self::AccountId>;
    type CollectionId: Parameter + Member + MaybeSerializeDeserialize + Copy + Default + TypeInfo + IsType<<<Self as Config>::NonFungibles as Inspect<<Self as Config>::AccountId>>::CollectionId> + MaxEncodedLen;
    type ItemId: Parameter + Member + MaybeSerializeDeserialize + Copy + TypeInfo + From<u128> + IsType<<<Self as Config>::NonFungibles as Inspect<<Self as Config>::AccountId>>::ItemId> + MaxEncodedLen;
    type PalletId: Get<PalletId>;
}

Configuration trait of this pallet.

Implement this type for a runtime in order to customize this pallet.

Required Associated Types

type RuntimeEvent: From<Event<Self>> + IsType<<Self as Config>::RuntimeEvent>

type WeightInfo: WeightInfo

type Fungibles: Mutate<Self::AccountId>

Fungibles implements fungibles::Mutate, granting us a way of charging the buyer of an NFT the respective asking price.

type NonFungibles: Transfer<Self::AccountId>

The NonFungibles trait impl that can transfer and inspect NFTs.

type CollectionId: Parameter + Member + MaybeSerializeDeserialize + Copy + Default + TypeInfo + IsType<<<Self as Config>::NonFungibles as Inspect<<Self as Config>::AccountId>>::CollectionId> + MaxEncodedLen

The NFT CollectionId type

type ItemId: Parameter + Member + MaybeSerializeDeserialize + Copy + TypeInfo + From<u128> + IsType<<<Self as Config>::NonFungibles as Inspect<<Self as Config>::AccountId>>::ItemId> + MaxEncodedLen

The NFT ItemId type

type PalletId: Get<PalletId>

The Id of this pallet